Move cipher.rs to new crypto module
The intention is to delineate interfaces to be implemented by
pluggable crypto providers.
